+++ packages/tests/tests/ephemeral.node.spec.ts
@@ -1,17 +1,24 @@
 import { bytesToUtf8, utf8ToBytes } from "@waku/byte-utils";
-import { waitForRemotePeer } from "@waku/core/lib/wait_for_remote_peer";
-import { DecoderV0, EncoderV0 } from "@waku/core/lib/waku_message/version_0";
-import { createLightNode } from "@waku/create";
-import { DecodedMessage, Protocols, WakuLight } from "@waku/interfaces";
 import {
-  AsymDecoder,
-  AsymEncoder,
+  createDecoder,
+  createEncoder,
+  DecodedMessage,
+  waitForRemotePeer,
+} from "@waku/core";
+import { createLightNode } from "@waku/create";
+import { Protocols } from "@waku/interfaces";
+import type { WakuLight } from "@waku/interfaces";
+import {
+  createDecoder as eciesDecoder,
+  createEncoder as eciesEncoder,
   generatePrivateKey,
-  generateSymmetricKey,
   getPublicKey,
-  SymDecoder,
-  SymEncoder,
-} from "@waku/message-encryption";
+} from "@waku/message-encryption/ecies";
+import {
+  generateSymmetricKey,
+  createDecoder as symDecoder,
+  createEncoder as symEncoder,
+} from "@waku/message-encryption/symmetric";
 import { expect } from "chai";
 import debug from "debug";
 
@@ -26,8 +33,8 @@ import {
 const log = debug("waku:test:ephemeral");
 
 const TestContentTopic = "/test/1/ephemeral/utf8";
-const TestEncoder = new EncoderV0(TestContentTopic);
-const TestDecoder = new DecoderV0(TestContentTopic);
+const TestEncoder = createEncoder(TestContentTopic);
+const TestDecoder = createDecoder(TestContentTopic);
 
 describe("Waku Message Ephemeral field", () => {
   let waku: WakuLight;
@@ -79,17 +86,17 @@ describe("Waku Message Ephemeral field", () => {
     const AsymContentTopic = "/test/1/ephemeral-asym/utf8";
     const SymContentTopic = "/test/1/ephemeral-sym/utf8";
 
-    const asymEncoder = new AsymEncoder(
+    const asymEncoder = eciesEncoder(
       AsymContentTopic,
       publicKey,
       undefined,
       true
     );
-    const symEncoder = new SymEncoder(SymContentTopic, symKey, undefined, true);
-    const clearEncoder = new EncoderV0(TestContentTopic, true);
+    const symEncoder = eciesEncoder(SymContentTopic, symKey, undefined, true);
+    const clearEncoder = createEncoder(TestContentTopic, true);
 
-    const asymDecoder = new AsymDecoder(AsymContentTopic, privateKey);
-    const symDecoder = new SymDecoder(SymContentTopic, symKey);
+    const asymDecoder = eciesDecoder(AsymContentTopic, privateKey);
+    const symDecoder = eciesDecoder(SymContentTopic, symKey);
 
     const [waku1, waku2, nimWakuMultiaddr] = await Promise.all([
       createLightNode({
@@ -146,7 +153,7 @@ describe("Waku Message Ephemeral field", () => {
   it("Ephemeral field is preserved - encoder v0", async function () {
     this.timeout(10000);
 
-    const ephemeralEncoder = new EncoderV0(TestContentTopic, true);
+    const ephemeralEncoder = createEncoder(TestContentTopic, true);
 
     const messages: DecodedMessage[] = [];
     const callback = (msg: DecodedMessage): void => {
@@ -186,14 +193,14 @@ describe("Waku Message Ephemeral field", () => {
 
     const symKey = generateSymmetricKey();
 
-    const ephemeralEncoder = new SymEncoder(
+    const ephemeralEncoder = symEncoder(
       TestContentTopic,
       symKey,
       undefined,
       true
     );
-    const encoder = new SymEncoder(TestContentTopic, symKey);
-    const decoder = new SymDecoder(TestContentTopic, symKey);
+    const encoder = symEncoder(TestContentTopic, symKey);
+    const decoder = symDecoder(TestContentTopic, symKey);
 
     const messages: DecodedMessage[] = [];
     const callback = (msg: DecodedMessage): void => {
@@ -234,14 +241,14 @@ describe("Waku Message Ephemeral field", () => {
     const privKey = generatePrivateKey();
     const pubKey = getPublicKey(privKey);
 
-    const ephemeralEncoder = new AsymEncoder(
+    const ephemeralEncoder = eciesEncoder(
       TestContentTopic,
       pubKey,
       undefined,
       true
     );
-    const encoder = new AsymEncoder(TestContentTopic, pubKey);
-    const decoder = new AsymDecoder(TestContentTopic, privKey);
+    const encoder = eciesEncoder(TestContentTopic, pubKey);
+    const decoder = eciesDecoder(TestContentTopic, privKey);
 
     const messages: DecodedMessage[] = [];
     const callback = (msg: DecodedMessage): void => {